requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
Appium-Python-Client>=3.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

